refactor: Start worker with server to ensure it runs
The background worker was not starting because it was a separate process that was not being initiated.

This change refactors the worker logic into a `worker-manager.js` file.

The main `server.js` now imports and starts the worker, ensuring background jobs are always processed when the server is running.

This also resolves the issue of not seeing worker logs, as they will now appear in the main server terminal output.

The original `worker.js` is kept for compatibility with the Procfile and now uses the new manager.
